requirements.txt
fastapi>=0.111.1
uvicorn[standard]>=0.29.0
gunicorn>=21.2.0

pinecone>=3.0.0

langchain>=0.1.20
langchain-community>=0.0.36
langchain-text-splitters>=0.0.1

sentence-transformers>=2.6.1

google-genai>=0.3.0

pypdf>=3.12.0
python-dotenv>=1.0.1
requests>=2.32.0
tqdm>=4.66.1
